Upstream

+++ src/com/mcml/space/optimize/FireLimitor.java
@@ -13,28 +13,28 @@ import com.mcml.space.core.VLagger;
 
 public class FireLimitor implements Listener {
 
-	private final static HashMap<Chunk, Long> ChunkChecked = new HashMap<Chunk, Long>();
+    private final static HashMap<Chunk, Long> ChunkChecked = new HashMap<Chunk, Long>();
 
-	@EventHandler
-	public void WhenFireLimitor(BlockIgniteEvent event) {
-		if (VLagger.FireLimitorenable == true) {
-			if (event.getCause() == IgniteCause.FLINT_AND_STEEL) {
-				return;
-			}
-			Block block = event.getBlock();
-			if (CheckFast(block.getChunk())) {
-				event.setCancelled(true);
-			} else {
-				ChunkChecked.put(block.getChunk(), System.currentTimeMillis());
-			}
-		}
-	}
+    @EventHandler
+    public void WhenFireLimitor(BlockIgniteEvent event) {
+        if (VLagger.FireLimitorenable == true) {
+            if (event.getCause() == IgniteCause.FLINT_AND_STEEL) {
+                return;
+            }
+            Block block = event.getBlock();
+            if (CheckFast(block.getChunk())) {
+                event.setCancelled(true);
+            } else {
+                ChunkChecked.put(block.getChunk(), System.currentTimeMillis());
+            }
+        }
+    }
 
-	private static boolean CheckFast(Chunk chunk) {
-		if (ChunkChecked.containsKey(chunk)) {
-			return (((Long) ChunkChecked.get(chunk)).longValue() + VLagger.FireLimitorPeriod > System
-					.currentTimeMillis());
-		}
-		return false;
-	}
+    private static boolean CheckFast(Chunk chunk) {
+        if (ChunkChecked.containsKey(chunk)) {
+            return (ChunkChecked.get(chunk).longValue() + VLagger.FireLimitorPeriod > System
+                    .currentTimeMillis());
+        }
+        return false;
+    }
 }
